+++ pkg/odo/cli/pipelines/bootstrap.go
@@ -9,7 +9,6 @@ import (
 	"github.com/openshift/odo/pkg/odo/genericclioptions"
 	"github.com/openshift/odo/pkg/pipelines"
 	"github.com/openshift/odo/pkg/pipelines/ioutils"
-	"github.com/openshift/odo/pkg/util"
 	"github.com/spf13/cobra"
 
 	ktemplates "k8s.io/kubectl/pkg/util/templates"
@@ -64,20 +63,7 @@ func (io *BootstrapParameters) Validate() error {
 	if len(removeEmptyStrings(strings.Split(gr.Path, "/"))) != 2 {
 		return fmt.Errorf("repo must be org/repo: %s", strings.Trim(gr.Path, ".git"))
 	}
-	if io.GitOpsWebhookSecret == "" {
-		gitopsSecret, err := util.GenerateSecureString(20)
-		if err != nil {
-			return err
-		}
-		io.GitOpsWebhookSecret = gitopsSecret
-	}
-	if io.AppWebhookSecret == "" {
-		appSecret, err := util.GenerateSecureString(20)
-		if err != nil {
-			return err
-		}
-		io.AppWebhookSecret = appSecret
-	}
+
 	return nil
 }
