+++ pkg/pipelines/secret/generate_secret.go
@@ -1,4 +1,4 @@
-package util
+package secret
 
 import (
 	"crypto/rand"
