+++ pkg/pipelines/secret/generate_secret_test.go
@@ -1,4 +1,4 @@
-package util
+package secret
 
 import (
 	"testing"
